Fixed merge conflict due to renaming of utils.js to utils.ts on the typescript-rework branch
